Fix policies and update PolicyRepositoryProvider
