+++ api/src/pdc/services/policy/engine/policies/ATMB202305.spec.ts
@@ -2,7 +2,7 @@ import test from 'ava';
 import { v4 } from 'uuid';
 import { OperatorsEnum } from '../../interfaces';
 import { makeProcessHelper } from '../tests/macro';
-import { Atmb as Handler } from './Atmb';
+import { ATMB202305 as Handler } from './ATMB202305';
 
 const defaultPosition = {
   arr: '74278',
