+++ api/src/pdc/services/policy/engine/policies/Cannes2024.spec.ts
@@ -2,17 +2,17 @@ import test from 'ava';
 import { v4 } from 'uuid';
 import { OperatorsEnum } from '../../interfaces';
 import { makeProcessHelper } from '../tests/macro';
-import { Smt as Handler } from './Smt';
+import { Cannes2024 as Handler } from './Cannes2024';
 
 const defaultPosition = {
-  arr: '37261',
-  com: '37261',
-  aom: '200085108',
-  epci: '243700754',
-  dep: '37',
-  reg: '24',
+  arr: '74278',
+  com: '74278',
+  aom: '200033116',
+  epci: '200033116',
+  dep: '74',
+  reg: '84',
   country: 'XXXXX',
-  reseau: '96',
+  reseau: '142',
 };
 const defaultLat = 48.72565703413325;
 const defaultLon = 2.261827843187402;
@@ -23,11 +23,11 @@ const defaultCarpool = {
   passenger_identity_uuid: v4(),
   driver_identity_uuid: v4(),
   operator_uuid: OperatorsEnum.KLAXIT,
-  operator_class: 'C',
+  operator_class: 'B',
   passenger_is_over_18: true,
   passenger_has_travel_pass: true,
   driver_has_travel_pass: true,
-  datetime: new Date('2019-01-15'),
+  datetime: new Date('2024-05-15'),
   seats: 1,
   duration: 600,
   distance: 5_000,
@@ -49,7 +49,7 @@ test(
   process,
   {
     policy: { handler: Handler.id },
-    carpool: [{ operator_uuid: 'not in list' }, { distance: 100 }, { distance: 200_000 }, { operator_class: 'A' }],
+    carpool: [{ operator_uuid: 'not in list' }, { distance: 100 }, { distance: 80_001 }, { operator_class: 'A' }],
     meta: [],
   },
   { incentive: [0, 0, 0, 0], meta: [] },
@@ -63,69 +63,52 @@ test(
     carpool: [
       { distance: 5_000, driver_identity_uuid: 'one' },
       { distance: 5_000, seats: 2, driver_identity_uuid: 'one' },
-      { distance: 25_000, driver_identity_uuid: 'two' },
-      { distance: 25_000, driver_identity_uuid: 'two', datetime: new Date('2022-03-28') },
-      { distance: 55_000, driver_identity_uuid: 'two' },
+      { distance: 30_000, driver_identity_uuid: 'one' },
+      { distance: 60_000, driver_identity_uuid: 'one' },
     ],
     meta: [],
   },
   {
-    incentive: [200, 400, 250, 250, 400],
+    incentive: [150, 300, 300, 300],
     meta: [
       {
+        key: 'max_amount_restriction.0-one.month.4-2024',
+        value: 1050,
+      },
+      {
         key: 'max_amount_restriction.global.campaign.global',
-        value: 1500,
+        value: 1050,
       },
     ],
   },
 );
 
 test(
-  'should work with global limits',
-  process,
-  {
-    policy: { handler: Handler.id, max_amount: 40_000_00 },
-    carpool: [{ distance: 5_000, driver_identity_uuid: 'one' }],
-    meta: [
-      {
-        key: 'max_amount_restriction.global.campaign.global',
-        value: 39_999_50,
-      },
-    ],
-  },
-  {
-    incentive: [50],
-    meta: [
-      {
-        key: 'max_amount_restriction.global.campaign.global',
-        value: 40_000_00,
-      },
-    ],
-  },
-);
-
-test(
-  'should work with day limits',
+  'should work with driver month limits 150',
   process,
   {
     policy: { handler: Handler.id },
     carpool: [
       { distance: 5_000, driver_identity_uuid: 'one' },
       { distance: 5_000, driver_identity_uuid: 'one' },
-      { distance: 5_000, driver_identity_uuid: 'one' },
-      { distance: 5_000, driver_identity_uuid: 'one' },
-      { distance: 5_000, driver_identity_uuid: 'one' },
-      { distance: 5_000, driver_identity_uuid: 'one' },
-      { distance: 5_000, driver_identity_uuid: 'one' },
     ],
-    meta: [],
-  },
-  {
-    incentive: [200, 200, 200, 200, 200, 200, 0],
     meta: [
       {
+        key: 'max_amount_restriction.0-one.month.4-2024',
+        value: 148_50,
+      },
+    ],
+  },
+  {
+    incentive: [150, 0],
+    meta: [
+      {
+        key: 'max_amount_restriction.0-one.month.4-2024',
+        value: 150_00,
+      },
+      {
         key: 'max_amount_restriction.global.campaign.global',
-        value: 12_00,
+        value: 150,
       },
     ],
   },
