+++ api/src/pdc/services/policy/engine/policies/PaysBasque20232024.ts
@@ -7,6 +7,7 @@ import {
   StatelessContextInterface,
 } from '../../interfaces';
 import {
+  isAdultOrThrow,
   isOperatorClassOrThrow,
   isOperatorOrThrow,
   LimitTargetEnum,
@@ -19,31 +20,23 @@ import {
   watchForPersonMaxTripByDay,
 } from '../helpers';
 import { AbstractPolicyHandler } from './AbstractPolicyHandler';
-import { description } from './Smt2023.html';
-import { dateWithTz, today } from '../../helpers';
+import { description } from './PaysBasque20232024.html';
 
-// Politique du Syndicat des Mobilités de Touraine
-export const Smt2023: PolicyHandlerStaticInterface = class
+// Pays Basque Adour
+export const PaysBasque20232024: PolicyHandlerStaticInterface = class
   extends AbstractPolicyHandler
   implements PolicyHandlerInterface
 {
-  static readonly id = 'smt_2023';
-  protected operators = [OperatorsEnum.KLAXIT];
-  protected new_operators = [OperatorsEnum.KLAXIT, OperatorsEnum.BLABLACAR_DAILY];
-
-  protected relaunch_update_date = new Date('15/02/2024');
-
-  protected slices: RunnableSlices = [
-    { start: 2_000, end: 20_000, fn: (ctx: StatelessContextInterface) => perSeat(ctx, 200) },
-    {
-      start: 20_000,
-      end: 40_000,
-      fn: (ctx: StatelessContextInterface) => perSeat(ctx, perKm(ctx, { amount: 10, offset: 20_000, limit: 40_000 })),
-    },
+  static readonly id = 'pays_basque_2023';
+  protected operators = [
+    OperatorsEnum.KLAXIT,
+    OperatorsEnum.MOBICOOP,
+    OperatorsEnum.BLABLACAR_DAILY,
+    OperatorsEnum.KAROS,
   ];
-
-  protected slices_after_april: RunnableSlices = [
-    { start: 2_000, end: 20_000, fn: (ctx: StatelessContextInterface) => perSeat(ctx, 200) },
+  protected operators_for_2024 = [OperatorsEnum.KLAXIT, OperatorsEnum.BLABLACAR_DAILY, OperatorsEnum.KAROS];
+  protected slices: RunnableSlices = [
+    { start: 5_000, end: 20_000, fn: (ctx: StatelessContextInterface) => perSeat(ctx, 200) },
     {
       start: 20_000,
       end: 30_000,
@@ -51,23 +44,29 @@ export const Smt2023: PolicyHandlerStaticInterface = class
     },
   ];
 
+  private readonly LAST_DAY_OF_YEAR_DATE = new Date('2023-12-31');
+
   constructor(public max_amount: number) {
     super();
     this.limits = [
       ['A34719E4-DCA0-78E6-38E4-701631B106C2', 6, watchForPersonMaxTripByDay, LimitTargetEnum.Driver],
-      // 120€ to 150€ from 15/02/2024
       ['ECDE3CD4-96FF-C9D2-BA88-45754205A798', 150_00, watchForPersonMaxAmountByMonth, LimitTargetEnum.Driver],
       ['B15AD9E9-BF92-70FA-E8F1-B526D1BB6D4F', this.max_amount, watchForGlobalMaxAmount],
     ];
   }
 
   protected processExclusion(ctx: StatelessContextInterface) {
-    isOperatorOrThrow(ctx, this.isAfter15February2024(ctx.carpool.datetime) ? this.new_operators : this.operators);
-    isOperatorClassOrThrow(ctx, ['B', 'C']);
-    onDistanceRangeOrThrow(
-      ctx,
-      this.isAfter15April2023(ctx.carpool.datetime) ? { min: 2_000, max: 30_001 } : { min: 2_000 },
-    );
+    isOperatorOrThrow(ctx, this.isAfter2023(ctx) ? this.operators_for_2024 : this.operators);
+    onDistanceRangeOrThrow(ctx, {
+      min: 5_000,
+      max: this.isAfter2023(ctx) ? 80_000 : 100_000,
+    });
+    isOperatorClassOrThrow(ctx, ['C']);
+    isAdultOrThrow(ctx);
+  }
+
+  private isAfter2023(ctx: StatelessContextInterface) {
+    return ctx.carpool.datetime > this.LAST_DAY_OF_YEAR_DATE;
   }
 
   processStateless(ctx: StatelessContextInterface): void {
@@ -76,7 +75,7 @@ export const Smt2023: PolicyHandlerStaticInterface = class
 
     // Par kilomètre
     let amount = 0;
-    for (const { start, fn } of this.getSlices()) {
+    for (const { start, fn } of this.slices) {
       if (onDistanceRange(ctx, { min: start })) {
         amount += fn(ctx);
       }
@@ -85,23 +84,11 @@ export const Smt2023: PolicyHandlerStaticInterface = class
     ctx.incentive.set(amount);
   }
 
-  private getSlices() {
-    return this.isAfter15April2023(today()) ? this.slices_after_april : this.slices;
-  }
-
-  private isAfter15February2024(date: Date): boolean {
-    return date >= dateWithTz(new Date('2024-02-15'));
-  }
-
-  private isAfter15April2023(date: Date): boolean {
-    return date >= dateWithTz(new Date('2023-04-15'));
-  }
-
   params(): PolicyHandlerParamsInterface {
     return {
       tz: 'Europe/Paris',
-      slices: this.getSlices(),
-      operators: this.new_operators,
+      slices: this.slices,
+      operators: this.operators_for_2024,
       limits: {
         glob: this.max_amount,
       },
