+++ api/src/pdc/services/policy/engine/policies/PaysDeLaLoire2021.spec.ts
@@ -2,7 +2,7 @@ import test from 'ava';
 import { v4 } from 'uuid';
 import { OperatorsEnum } from '../../interfaces';
 import { makeProcessHelper } from '../tests/macro';
-import { Pdll2023 as Handler } from './Pdll2023';
+import { PaysDeLaLoire2021 as Handler } from './PaysDeLaLoire2021';
 
 const defaultPosition = {
   arr: '85047',
@@ -27,7 +27,7 @@ const defaultCarpool = {
   passenger_is_over_18: true,
   passenger_has_travel_pass: true,
   driver_has_travel_pass: true,
-  datetime: new Date('2023-04-15'),
+  datetime: new Date('2019-01-15'),
   seats: 1,
   duration: 600,
   distance: 5_000,
@@ -50,6 +50,7 @@ test(
   {
     policy: { handler: Handler.id },
     carpool: [
+      { operator_uuid: OperatorsEnum.ATCHOUM },
       { distance: 100 },
       { operator_class: 'A' },
       { start: { ...defaultPosition, aom: '244900015' }, end: { ...defaultPosition, aom: '244900015' } },
@@ -58,12 +59,10 @@ test(
       { start: { ...defaultPosition, aom: '200071678' }, end: { ...defaultPosition, aom: '200071678' } },
       { start: { ...defaultPosition, reg: '11' } },
       { end: { ...defaultPosition, reg: '11' } },
-      { distance: 90_000 },
-      { passenger_is_over_18: false },
     ],
     meta: [],
   },
-  { incentive: [0, 0, 0, 0, 0, 0, 0, 0, 0, 0], meta: [] },
+  { incentive: [0, 0, 0, 0, 0, 0, 0, 0, 0], meta: [] },
 );
 
 test(
@@ -72,37 +71,20 @@ test(
   {
     policy: { handler: Handler.id },
     carpool: [
-      { distance: 1_000, driver_identity_uuid: 'one' },
       { distance: 5_000, driver_identity_uuid: 'one' },
       { distance: 5_000, seats: 2, driver_identity_uuid: 'one' },
-      { distance: 20_000, driver_identity_uuid: 'two' },
       { distance: 25_000, driver_identity_uuid: 'two' },
       { distance: 25_000, driver_identity_uuid: 'two', datetime: new Date('2022-03-28') },
-      { distance: 40_000, driver_identity_uuid: 'two' },
       { distance: 55_000, driver_identity_uuid: 'two' },
-      { distance: 80_000, driver_identity_uuid: 'two' },
-      { distance: 90_000, driver_identity_uuid: 'two' },
     ],
     meta: [],
   },
   {
-    incentive: [0, 100, 200, 100, 150, 150, 300, 300, 300, 0],
+    incentive: [200, 400, 250, 250, 500],
     meta: [
       {
         key: 'max_amount_restriction.global.campaign.global',
         value: 1600,
-      },
-      {
-        key: 'max_amount_restriction.0-one.month.3-2023',
-        value: 300,
-      },
-      {
-        key: 'max_amount_restriction.0-two.month.3-2023',
-        value: 1150,
-      },
-      {
-        key: 'max_amount_restriction.0-two.month.2-2022',
-        value: 150,
       },
     ],
   },
@@ -112,12 +94,12 @@ test(
   'should work with global limits',
   process,
   {
-    policy: { handler: Handler.id, max_amount: 500_000_00 },
+    policy: { handler: Handler.id, max_amount: 2_000_000_00 },
     carpool: [{ distance: 5_000, driver_identity_uuid: 'one' }],
     meta: [
       {
         key: 'max_amount_restriction.global.campaign.global',
-        value: 499_999_50,
+        value: 1_999_999_50,
       },
     ],
   },
@@ -126,11 +108,7 @@ test(
     meta: [
       {
         key: 'max_amount_restriction.global.campaign.global',
-        value: 500_000_00,
-      },
-      {
-        key: 'max_amount_restriction.0-one.month.3-2023',
-        value: 100,
+        value: 2_000_000_00,
       },
     ],
   },
@@ -153,15 +131,11 @@ test(
     meta: [],
   },
   {
-    incentive: [100, 100, 100, 100, 100, 100, 0],
+    incentive: [200, 200, 200, 200, 200, 200, 0],
     meta: [
       {
         key: 'max_amount_restriction.global.campaign.global',
-        value: 600,
-      },
-      {
-        key: 'max_amount_restriction.0-one.month.3-2023',
-        value: 600,
+        value: 1200,
       },
     ],
   },
