+++ api/src/pdc/services/policy/engine/policies/SMT2023.spec.ts
@@ -2,17 +2,17 @@ import test from 'ava';
 import { v4 } from 'uuid';
 import { OperatorsEnum } from '../../interfaces';
 import { makeProcessHelper } from '../tests/macro';
-import { PaysBasque as Handler } from './PaysBasque';
+import { SMT2023 as Handler } from './SMT2023';
 
 const defaultPosition = {
-  arr: '64155',
-  com: '64155',
-  aom: '256401605',
-  epci: '200067106',
-  dep: '64',
-  reg: '75',
+  arr: '37050',
+  com: '37050',
+  aom: '200085108',
+  epci: '243700754',
+  dep: '37',
+  reg: '24',
   country: 'XXXXX',
-  reseau: '15',
+  reseau: '96',
 };
 const defaultLat = 48.72565703413325;
 const defaultLon = 2.261827843187402;
@@ -23,14 +23,14 @@ const defaultCarpool = {
   passenger_identity_uuid: v4(),
   driver_identity_uuid: v4(),
   operator_uuid: OperatorsEnum.KLAXIT,
-  operator_class: 'C',
+  operator_class: 'B',
   passenger_is_over_18: true,
   passenger_has_travel_pass: true,
   driver_has_travel_pass: true,
   datetime: new Date('2023-04-15'),
   seats: 1,
   duration: 600,
-  distance: 6_000,
+  distance: 5_000,
   cost: 20,
   driver_payment: 20,
   passenger_payment: 20,
@@ -69,13 +69,11 @@ test(
           reg: '24',
         },
       },
-      { passenger_is_over_18: false },
-      { distance: 90_000, datetime: new Date('2024-01-01') },
-      { datetime: new Date('2024-01-01'), operator_uuid: OperatorsEnum.MOBICOOP },
+      { distance: 31_000, datetime: new Date('2023-04-17') },
     ],
     meta: [],
   },
-  { incentive: [0, 0, 0, 0, 0, 0, 0], meta: [] },
+  { incentive: [0, 0, 0, 0, 0], meta: [] },
 );
 
 test(
@@ -89,28 +87,29 @@ test(
       { distance: 25_000, driver_identity_uuid: 'one' },
       { distance: 25_000, seats: 2, driver_identity_uuid: 'one' },
       {
-        distance: 35_000,
+        distance: 29_000,
         driver_identity_uuid: 'one',
+        end: {
+          arr: '37109',
+          com: '37109',
+          aom: '200085108',
+          epci: '243700754',
+          reg: '24',
+        },
       },
-      { distance: 90_000, datetime: new Date('2023-12-31'), driver_identity_uuid: 'one' },
-      { distance: 80_000, driver_identity_uuid: 'one' },
     ],
     meta: [],
   },
   {
-    incentive: [200, 400, 250, 500, 300, 300, 300],
+    incentive: [200, 400, 250, 500, 290],
     meta: [
       {
         key: 'max_amount_restriction.0-one.month.3-2023',
-        value: 1950,
+        value: 1640,
       },
       {
         key: 'max_amount_restriction.global.campaign.global',
-        value: 2250,
-      },
-      {
-        key: 'max_amount_restriction.0-one.month.11-2023',
-        value: 300,
+        value: 1640,
       },
     ],
   },
@@ -125,12 +124,12 @@ test(
     meta: [
       {
         key: 'max_amount_restriction.global.campaign.global',
-        value: 49_998_00,
+        value: 59_999_50,
       },
     ],
   },
   {
-    incentive: [200],
+    incentive: [50],
     meta: [
       {
         key: 'max_amount_restriction.0-one.month.3-2023',
@@ -138,7 +137,7 @@ test(
       },
       {
         key: 'max_amount_restriction.global.campaign.global',
-        value: 50_000_00,
+        value: 60_000_00,
       },
     ],
   },
