+++ api/src/pdc/services/policy/engine/policies/Vitre2023.html.ts
@@ -4,16 +4,19 @@ export const description = `<p _ngcontent-pmm-c231="" id="summary" class="campai
 <p>Cette campagne est limitée à
   <b>361 000 euros </b>.
 </p>
-<p>Les <b> passagers </b> effectuant un trajet d'au moins 2 km avec pour origine ET destination dans la l'AOM Vitré Communauté 
-sont incités.
+<p>
+  Les <b> passagers </b> effectuant un trajet d'au moins 2 km avec pour origine ET destination 
+  dans la l'AOM Vitré Communauté sont incités.
 </p>
-<p>Les <b> conducteurs et passagers </b> effectuant un trajet d'au moins 2 km sont incités selon les règles suivantes avant le 17 Juillet 2023:
+<p>Les <b> conducteurs et passagers </b> effectuant un trajet d'au moins 2 km sont incités selon les 
+règles suivantes avant le 17 Juillet 2023:
 </p>
 <ul>
   <li><b>De 2 à 15 km : 1.5 euros par trajet par passager </b></li>
   <li><b>De 15 à 30 km : 0.1 euro par trajet par km par passager </b></li>
 </ul>
-<p>Les <b> conducteurs et passagers </b> effectuant un trajet d'au moins 2 km sont incités selon les règles suivantes après le 17 Juillet 2023 (inclus):
+<p>Les <b> conducteurs et passagers </b> effectuant un trajet d'au moins 2 km sont incités selon 
+les règles suivantes après le 17 Juillet 2023 (inclus):
 </p>
 <ul>
   <li><b>De 2 à 10 km : 1 euro par trajet par passager </b></li>
