+++ api/src/pdc/services/policy/engine/policies/Vitre2023.spec.ts
@@ -2,7 +2,7 @@ import test from 'ava';
 import { v4 } from 'uuid';
 import { OperatorsEnum } from '../../interfaces';
 import { makeProcessHelper } from '../tests/macro';
-import { Vitre as Handler } from './Vitre';
+import { Vitre2023 as Handler } from './Vitre2023';
 
 const defaultPosition = {
   arr: '35361',
